Fixed some field types, and imports in Tutor/Student.py

+++ backend/api/models/Student.py
@@ -1,6 +1,6 @@
 from django.db import models
 from api.models.User import User
-from backend.api.models.TutorSession import TutorSession
+from api.models.TutorSession import TutorSession
 
 class Student(models.Model):
 
@@ -20,7 +20,7 @@ class Student(models.Model):
   #PK
   studentID = models.PositiveSmallIntegerField(primary_key=True, blank=False, null=False)
   tutorSessionID= models.foreignKey(TutorSession, on_delete=models.CASCADE,blank=True, null=True)
-  tutorSessHours= models.IntegerField(default=0)
+  tutorSessHours= models.PositiveSmallIntegerField(default=0)
